\name{BACprior-package}
\alias{BACprior-package}
\alias{BACprior}
\docType{package}
\title{
Sensitivity of the Bayesian Adjustment for Confounding (BAC) algorithm to the choice of hyperparameter omega}
\description{
The BACprior package contains only one function at the moment whose output allows the user to evaluate the sensitivity of BAC to the omega value chosen 
in the covariate inclusion indicators' (of the outcome and exposure models) prior distribution. }
\details{
\tabular{ll}{
Package: \tab BACprior\cr
Type: \tab Package\cr
Version: \tab 1.2.1\cr
Date: \tab 2013-02-15\cr
License: \tab GPL (>=2) \cr
}
}
\author{
Denis jf Talbot, Geneviève Lefebvre, Juli Atherton.

\email{talbot.denis@courrier.uqam.ca}}
\references{

%Brookhart, M.A., van der Lan, M.J. (2006). \emph{A semiparametric model selection criterion with applications to the marginal structural model}, Computational Statistics & Data Analysis, 50, 475-498.

Hoeting, J., D. Madigan, A. Raftery, and C. Volinsky (1999). \emph{Bayesian model averaging : A tutorial}, Statistical Science, 16, 382-417.

Lefebvre, G., Atherton, J., Talbot, D. (2013). \emph{A note on the effect of the prior distribution in the Bayesian Adjustment for Confounding algorithm}, Submitted.

Wang, C., Parmigiani, G., Dominici, F. (2012). \emph{Bayesian effect estimation accounting for adjustment uncertainty}, Biometrics DOI: 10.1111/j.1541-0420.2011.01731.x.

%Wang, C., Parmigiani, G., Dominici, F. (2012b). \emph{Rejoinder: Bayesian effect estimation accounting for adjustment uncertainty}, Biometrics DOI: 10.1111/j.15410420.2011.01735.x.
}


