\name{BACprior-package}
\alias{BACprior-package}
\alias{BACprior}
\docType{package}
\title{
Choice of the Hyperparameter Omega in the Bayesian Adjustment for Confounding (BAC) Algorithm}
\description{
The BACprior package contains functions to help the user select the omega value appearing in the BAC prior distribution of the covariate inclusion indicators of the outcome and exposure models. }
\details{
\tabular{ll}{
Package: \tab BACprior\cr
Type: \tab Package\cr
Version: \tab 2.1.2\cr
Date: \tab 2025-08-28\cr
License: \tab GPL (>=2)
}
}
\author{
Denis Talbot, Genevieve Lefebvre, Juli Atherton.

Maintainer: Denis Talbot \email{denis.talbot@fmed.ulaval.ca}}
\references{
Brookhart, M.A., van der Lan, M.J. (2006). \emph{A semiparametric model selection criterion with applications to the marginal structural model}, Computational Statistics & Data Analysis, 50, 475-498.

Hoeting, J.A., Madigan D., Raftery, A.E., Volinsky C.T. (1999). \emph{Bayesian model averaging : A tutorial}, Statistical Science, 16, 382-417.

Lefebvre, G., Atherton, J., Talbot, D. (2014). \emph{The effect of the prior distribution in the Bayesian Adjustment for Confounding algorithm}, Computational Statistics & Data Analysis, 70, 227-240.

Wang, C., Parmigiani, G., Dominici, F. (2012). \emph{Bayesian effect estimation accounting for adjustment uncertainty}, Biometrics, 68 (3), 661-671.

%Wang, C., Parmigiani, G., Dominici, F. (2012b). \emph{Rejoinder: Bayesian effect estimation accounting for adjustment uncertainty}, Biometrics DOI: 10.1111/j.15410420.2011.01735.x.
}


